+++ src/components/sections/home/featuredArticles/index.tsx
@@ -1,40 +1,5 @@
-import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
-import { formatDateByLang } from '@/lib/utils';
-import {
-  FeaturedArticlesItemSchema,
-  FeaturedArticlesSchema,
-  LangSchema,
-} from '@/services/sanity/parser';
-import Link from 'next/link';
-import { PiClockBold } from 'react-icons/pi';
-
-type ItemProps = { data: FeaturedArticlesItemSchema; lang: LangSchema };
-
-function FeaturedArticlesItem({ data, lang }: ItemProps) {
-  return (
-    <Link href={`/${lang}/blog/${data.slug}`}>
-      <Card>
-        <CardHeader>
-          <CardTitle className="text-lg font-semibold">{data.title}</CardTitle>
-        </CardHeader>
-        <CardContent className="flex flex-col items-end justify-start w-full gap-2 sm:gap-5">
-          <p className="text-xs sm:text-base leading-6 sm:leading-7 dark:text-gray-300 w-full">
-            {data.summary}
-          </p>
-          <div className="flex justify-between w-full mt-4 sm:mt-0">
-            <span className="text-xs flex items-center gap-2">
-              <PiClockBold className="text-lg" />
-              {data.estimatedReadingTime.value}
-            </span>
-            <span className="text-xs">
-              {formatDateByLang(data.createdAt, lang)}
-            </span>
-          </div>
-        </CardContent>
-      </Card>
-    </Link>
-  );
-}
+import { ArticleItem } from '@/components/ui/article';
+import { FeaturedArticlesSchema, LangSchema } from '@/services/sanity/parser';
 
 type Props = { data: FeaturedArticlesSchema; lang: LangSchema };
 
@@ -47,7 +12,7 @@ export function FeaturedArticles({ data, lang }: Props) {
         </h2>
         <div className="flex flex-col justify-start items-start gap-5">
           {data.items.map((item) => (
-            <FeaturedArticlesItem key={item._key} data={item} lang={lang} />
+            <ArticleItem key={item._key} data={item} lang={lang} />
           ))}
         </div>
       </div>