+++ src/services/sanity/schemaTypes/pages/blogType.ts
@@ -1,0 +1,41 @@
+import { defineField, defineType } from 'sanity';
+
+export const blogPageType = defineType({
+  name: 'blog',
+  title: 'Blog',
+  type: 'document',
+  fields: [
+    defineField({
+      name: 'language',
+      type: 'string',
+      readOnly: true,
+      hidden: true,
+    }),
+    defineField({
+      name: 'meta',
+      title: 'Metadata',
+      type: 'meta',
+    }),
+    defineField({
+      name: 'title',
+      title: 'Title',
+      type: 'string',
+    }),
+    defineField({
+      name: 'description',
+      title: 'Description',
+      type: 'text',
+    }),
+  ],
+  preview: {
+    select: {
+      title: 'meta.title',
+    },
+    prepare({ title }) {
+      return {
+        title: title || 'Blog',
+        subtitle: '/[lang]/blog',
+      };
+    },
+  },
+});